feat: Refactor asset-store-effects to include user information in uploadAsset and updateDBWithName
Refactor the asset-store-effects module to include the user information in the uploadAsset and updateDBWithName functions. This change ensures that the username is included in the asset file name when uploading and updating the database. It uses the concatLatestFrom operator to get the user information from the auth store. This improvement enhances the accuracy and traceability of asset uploads and updates.

Co-authored-by: GursikhHajra <gursikhsinghhajra@gmail.com>