Asssets per user (#33)
* feat: Refactor asset-store-effects to include user information in uploadAsset and updateDBWithName

Refactor the asset-store-effects module to include the user information in the uploadAsset and updateDBWithName functions. This change ensures that the username is included in the asset file name when uploading and updating the database. It uses the concatLatestFrom operator to get the user information from the auth store. This improvement enhances the accuracy and traceability of asset uploads and updates.

Co-authored-by: GursikhHajra <gursikhsinghhajra@gmail.com>

* Refactor asset-store-effects to include user information in uploadAsset and updateDBWithName

* feat: Include username in search results

Update the PublicSearchComponent to include the username of the asset owner in the search results. This change adds a new property, "username", to the GetMediaListResponse interface and modifies the FlattenToGetMediaListResponse function to extract the username from the API response. The MediaManagementService now includes the username as a query parameter when making the API call to retrieve the assets. This enhancement improves the visibility and identification of asset owners in the search results.

Co-authored-by: GursikhHajra <gursikhsinghhajra@gmail.com>

* chore: Add download button to media-finder component

This commit adds a download button to the media-finder component's HTML template. The button allows users to download the displayed image. The button is styled with the "download_button" class in the component's SCSS file.

---------

Co-authored-by: GursikhHajra <gursikhsinghhajra@gmail.com>