UI 01 (#36)
* first

* 2

* media upload done

* feat: Add download button to media-finder component

* public view ux done

* its done

* chore: Refactor media-finder and uploader components

Refactor the media-finder and uploader components to improve their functionality and styling. This includes updating the CSS styles for the table in the media-finder component and the loading spinner in the uploader component. Additionally, the uploader component's HTML and SCSS files are refactored to improve the layout and file selection functionality.

* chore: Refactor media-finder and uploader components

* changed html

* Refactor auth-store.reducer and asset-store.reducer for improved functionality and code readability

* chore: Refactor media-finder and uploader components

Refactor the media-finder and uploader components to improve their functionality and styling. This includes updating the CSS styles for the table in the media-finder component and the loading spinner in the uploader component. Additionally, the uploader component's HTML and SCSS files are refactored to improve the layout and file selection functionality.

---------

Co-authored-by: gurjap <gurjaph@my.yorku.ca>