feat: Fix verify button functionality in media-finder component
Update the media-finder.component.html and media-finder.component.ts files to fix the functionality of the verify button. In the HTML file, change the binding of the click event to pass the entire asset object instead of just the asset name. In the TypeScript file, set the verifyLoading property to false after the verification process is complete. This improvement ensures that the verify button works correctly and provides a better user experience.