feat: Add lazy loading for images and confirm sign-up module
Add lazy loading for images in the app-routing.module.ts and header.component.html files. This improves the page load performance by loading images only when they are needed.

Also, add the confirm sign-up module with its routing in the app-routing.module.ts file. This module is responsible for handling the confirmation of user sign-up.

These changes enhance the functionality and usability of the application.