feat: Add scale control for image in media-finder.component.html
Add a scale control for the image in the media-finder.component.html file. This control allows users to adjust the scale of the image using a slider. The scale value is bound to the "scale" property in the component's TypeScript file. This enhancement improves the user experience by providing more control over the displayed image.